Two casts that stopped meaning anything when the steppers narrowed
cga_next_row returns uint16_t now, so `(uint16_t)cga_next_row(di)` casts
a value to the type it already has. Both sites went the same way, and the
`di` each was feeding it is a uint16_t too - net_pos and extra_pos are
both uint16_t fields, and what the local holds is a DI.

Identical to HEAD: the casts and the wider locals were describing a
truncation the types now carry.

+++ src/game.c
@@ -1231,10 +1231,10 @@ void game_input(void)
  * generalise over. */
 static void net_step(uint8_t reload)
 {
-    uint32_t di = global.net_pos;
+    uint16_t di = global.net_pos;
     g_vram[di & (CGA_SIZE - 1)] = 0;
     g_vram[(di + 1) & (CGA_SIZE - 1)] = 0;
-    global.net_pos = (uint16_t)cga_next_row(di);
+    global.net_pos = cga_next_row(di);
     global.net_timer = reload;
 }
 
@@ -1477,9 +1477,9 @@ frames:
                 /* One cell off the bar. `stosw` leaves di two on, and the
                  * two `dec di` put it back, so what follows is a plain step
                  * to the next scan line - the bar drains downwards. */
-                uint32_t di = global.extra_pos;
+                uint16_t di = global.extra_pos;
                 vram_setw(di, 0);
-                global.extra_pos = (uint16_t)(cga_next_row(di));
+                global.extra_pos = cga_next_row(di);
                 global.extra_timer = 400;
             }
             global.serve_timeout--;
